rpsql Documentation

Home 🏠

Functions postgresHasDefault postgresDefault postgresWaitForNotify postgresIsTransacting postgresImportLargeObject postgresExportLargeObject Postgres

Vignettes A Crash Course on PostgreSQL for R Users

postgresWaitForNotify {rpsql}	R Documentation

Wait for and return any notifications that return within timeout

Description

Once you subscribe to notifications with LISTEN, use this to wait for responses on each channel.

Usage

postgresWaitForNotify(conn, timeout = 1)

Arguments

conn 	 a [PqConnection-class] object, produced by [DBI::dbConnect()]

timeout 	 How long to wait, in seconds. Default 1

Value

If a notification was available, a list of:

channel

Name of channel

pid

PID of notifying server process

payload

Content of notification

If no notifications are available, return NULL